requirements.txt
attrs==19.3.0
boto3==1.14.55
botocore==1.17.55
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2; python_version >= "3.5"
chalice==1.19.0
chardet==3.0.4
click==7.1.2
docutils==0.15.2
enum-compat==0.0.3
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0; python_version >= "3.5"
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
jmespath==0.10.0
mypy-extensions==0.4.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
python-dateutil==2.8.1
pytz==2020.1
pyyaml==5.3.1
requests==2.24.0
rsa==4.6; python_version >= "3.5"
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
wheel==0.35.1
